Add wissen cell

+++ src/main/java/mod/maxbogomol/wizards_reborn/utils/RenderUtils.java
@@ -74,7 +74,7 @@ public class RenderUtils {
                     .setShaderState(new RenderStateShard.ShaderStateShard(WizardsRebornClient::getSpriteParticleShader))
                     .createCompositeState(false));
 
-    public static void renderItemModelInGui(ItemStack stack, int x, int y, int xSize, int ySize, int zSize) {
+    public static void renderItemModelInGui(ItemStack stack, int x, int y, float xSize, float ySize, float zSize) {
         BakedModel bakedmodel = Minecraft.getInstance().getItemRenderer().getModel(stack, (Level)null, (LivingEntity)null, 0);
 
         PoseStack posestack = RenderSystem.getModelViewStack();
